Update nNew.java
Each square in a grid represents a co-ordinate axis; the top-left square will have co-ordinates (50,50) and the bottom-right square will have co-ordinates (650, 650). For input arguments, user needs to put it in the form of 50s. Then, the algorithm will search for the fastest route to go from point A to point B while igniting the least number of squares possible.